add print

+++ application/views/surat/v_surat.php
@@ -19,9 +19,17 @@
 <div class="card mb-3">
     <div class="card-header">
     <i class="fas fa-table"></i>
-    Data user</div>
+    Data Surat</div>
     <div class="card-body">
-    <a href="<?php echo base_url() ?>surat/add" class="btn btn-primary"> Tambah Data</a><br> <br>
+    <?php
+     if($this->session->userdata('jabatan')!='staff'){
+
+     } else{ ?>
+         <a href="<?php echo base_url() ?>surat/add" class="btn btn-primary"> Tambah Data</a><br>
+   <?php }
+    ?>
+    <br> 
+    
     <div class="table-responsive">
     
         <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
@@ -30,7 +38,7 @@
         <th>no surat</th>
         <th>Tanggal Surat</th>
         <th>Perihal</th>
-        <th>Tujuan Undangan</th>
+        <th>Departemen</th>
         <th>Tanggal undangan</th>
         <th>Tempat Undangan</th>
         <th>Status</th>
@@ -46,10 +54,10 @@
         
         ?>
         <tr>
-        <td>005/<?php echo $d->id_surat ?>-SETDA/2019</td>
+        <td>005/<?php echo $d->id_surat ?>-<?php echo $this->session->userdata('departemen')?>/2019</td>
         <td><?php echo $d->tanggal_surat ?></td>
         <td><?php echo $d->perihal ?></td>
-        <td><?php echo $d->kepada ?></td>
+        <td><?php echo $d->departemen ?></td>
         <td><?php echo $d->tanggal_undangan ?></td>
         <td><?php echo $d->tempat ?></td>
         <td ><?php 
@@ -71,7 +79,7 @@
         <td><?php echo $d->tgl_update ?></td>
         </td>
         <td width="250px">
-            <?php if($d->status_approve==0||$d->status_approve==6) {?>
+            <?php if(($d->status_approve==0&&$this->session->userdata('jabatan')=='staff')||$d->status_approve==6) {?>
                 <a href="<?php echo base_url() ?>surat/edit/<?php echo $d->id_surat ?>" class="btn btn-primary"> Ubah</a>
                 <a href="<?php echo base_url() ?>surat/delete/<?php echo $d->id_surat ?>" onclick="return confirm('Yakin Hapus?')" class="btn btn-danger"> Hapus</a>
                 <a href="<?php echo base_url() ?>surat/detail/<?php echo $d->id_surat ?>" class="btn btn-success"> Detail</a>
